+++ packages/2d/src/lib/jsx-runtime.ts
@@ -15,7 +15,7 @@ export namespace JSX {
 }
 
 function isClassComponent(
-  // eslint-disable-next-line @typescript-eslint/ban-types
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
   fn: Function,
 ): fn is new (...args: unknown[]) => unknown {
   return !!fn.prototype?.isClass;